Make Latin module
